assessment/createAssessmentSchema: use bigint in reference

+++ server/repository/assessment/createAssessmentSchema.ts
@@ -34,7 +34,7 @@ create table ${schemaName}.section
     id            bigserial NOT NULL,
     uuid          uuid  default uuid_generate_v4(),
     props         jsonb default '{}'::jsonb,
-    assessment_id serial    not null references assessment (id),
+    assessment_id bigint   not null references assessment (id),
     PRIMARY KEY (id),
     unique(uuid)
 );
@@ -44,7 +44,7 @@ create table ${schemaName}.table_section
     id         bigserial NOT NULL,
     uuid       uuid  default uuid_generate_v4(),
     props      jsonb default '{}'::jsonb,
-    section_id serial    not null references ${schemaName}.section (id),
+    section_id bigint   not null references ${schemaName}.section (id),
     PRIMARY KEY (id),
     unique(uuid)
 );
@@ -55,7 +55,7 @@ create table ${schemaName}.table
     id               bigserial NOT NULL,
     uuid             uuid  default uuid_generate_v4(),
     props            jsonb default '{}'::jsonb,
-    table_section_id serial    not null references ${schemaName}.table_section (id),
+    table_section_id bigint   not null references ${schemaName}.table_section (id),
     PRIMARY KEY (id),
     unique(uuid)
 );
@@ -66,7 +66,7 @@ create table ${schemaName}.row
     id       bigserial NOT NULL,
     uuid     uuid  default uuid_generate_v4(),
     props    jsonb default '{}'::jsonb,
-    table_id serial    not null references ${schemaName}.table (id),
+    table_id bigint   not null references ${schemaName}.table (id),
     PRIMARY KEY (id),
     unique(uuid)
 );
@@ -76,7 +76,7 @@ create table ${schemaName}.col
     id     bigserial NOT NULL,
     uuid   uuid  default uuid_generate_v4(),
     props  jsonb default '{}'::jsonb,
-    row_id serial    not null references ${schemaName}.row (id),
+    row_id bigint   not null references ${schemaName}.row (id),
     PRIMARY KEY (id),
     unique(uuid)
 );
